+++ src/process/index.js
@@ -5,4 +5,5 @@
  * @memberof ran
  */
 export { default as BrownianMotion } from './brownian-motion'
+export { default as OrnsteinUhlenbeck } from './ornstein-uhlenbeck'
 export { default as Process } from './_process'